Remove symlink for check_ip_in_list and move it to HTTP API module
